google_ads_new/requirements_chat.txt
# Chat and AI Assistant Dependencies
# Install with: pip install -r requirements_chat.txt

# LangChain and OpenAI
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0

# OpenAI
openai>=1.0.0

# Vector embeddings and similarity search
sentence-transformers>=2.2.0
pgvector>=0.2.0

# Additional utilities
tiktoken>=0.5.0
numpy>=1.24.0

# Django REST Framework (if not already installed)
djangorestframework>=3.14.0

# For development and testing
python-dotenv>=1.0.0
